% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Analysis wrapper function}
\usage{
analysis(input, type = "binomial", .data = NULL)
}
\arguments{
\item{input}{list. Input function for all the analysis.}

\item{type}{character. Type of analysis to be ran (binomial (default),
normal. etc.).}

\item{.data}{NULL. stores the all the details, please do not fill it in.}
}
\value{
a list with results of the analysis of bayesian trial.
}
\description{
Wrapper function to analyze bayesian trials.
}
