% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialBACT.R
\name{BACTbinomial}
\alias{BACTbinomial}
\title{Complete binomial wrapper function}
\usage{
BACTbinomial(input, no_of_sim = 10000, .data = NULL)
}
\arguments{
\item{input}{list. Input function for all binomial_BACT.}

\item{no_of_sim}{numeric. Number of simulations to run}

\item{.data}{NULL. stores the proportion of control and treatment, please do not fill it in.}
}
\value{
a list with results of the clinical outcome.
}
\description{
Wrapper function for complete binomial bayesCT function to compute power and type 1 error.
}
