% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialBACT.R
\name{BACTbinomial_analysis}
\alias{BACTbinomial_analysis}
\title{Binomial analysis wrapper function}
\usage{
BACTbinomial_analysis(input, .data = NULL)
}
\arguments{
\item{input}{list. Input function for all binomial_analysis.}

\item{.data}{NULL. stores the binomial analysis details, please do not fill it in.}
}
\value{
a list with results of the analysis of bayesian trial.
}
\description{
Wrapper function for binomial bayesCT function to analyze bayesian trials.
}
