% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalBACT.R
\name{BACTnormal}
\alias{BACTnormal}
\title{Complete normal wrapper function}
\usage{
BACTnormal(input, .data = NULL, no_of_sim = 100)
}
\arguments{
\item{input}{list. Input function for all normalBACT.}

\item{.data}{NULL. stores the proportion of control and treatment, please do not fill it in.}

\item{no_of_sim}{numeric. Number of simulations to run.}
}
\value{
a list with results of the simulation (power and type I error).
}
\description{
Wrapper function for complete normal bayesCT function to compute power and type 1 error.
}
