% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{BACTnormal_analysis}
\alias{BACTnormal_analysis}
\title{Normal data analysis wrapper function}
\usage{
BACTnormal_analysis(input, .data = NULL)
}
\arguments{
\item{input}{list. Input function for all normal_analysis variables.}

\item{.data}{NULL. stores the normal analysis details, please do not fill it in.}
}
\value{
a list with results of the analysis of bayesian trial.
}
\description{
Wrapper function for normal bayesCT function to analyze bayesian trials.
}
