% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialBACT.R
\name{beta_prior}
\alias{beta_prior}
\title{Beta prior for for control and treatment group}
\usage{
beta_prior(prior = c(1, 1), .data = NULL)
}
\arguments{
\item{prior}{vector. Prior value of beta rate, \code{beta(a0, b0)}. The default is
set to \code{beta(1, 1)}.}

\item{.data}{NULL. stores the proportion of control and treatment, please do not fill it in.}
}
\value{
a list with vector of beta rate for the beta prior for treatment and control group.
}
\description{
Wrapper function for beta prior \code{beta(a0, b0)}.
}
\examples{
beta_prior(prior = c(1, 1))
}
