% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialBACT.R
\name{binomial_analysis}
\alias{binomial_analysis}
\title{Analyzing bayesian trial for binomial counts}
\usage{
binomial_analysis(data = data, y0_treatment = y0_treatment,
  N0_treatment = N0_treatment, y0_control = y0_control,
  N0_control = N0_control, alternative = "greater", N_impute = 100,
  number_mcmc = 10000, prob_ha = 0.95, futility_prob = 0.1,
  expected_success_prob = 0.9, prior = c(1, 1),
  discount_function = "identity", fix_alpha = TRUE, alpha_max = 1)
}
\arguments{
\item{data}{data frame. A data frame which provides patient id, treatment group, outcome
of the treatment and complete columns. An example file is available at
\code{data(binomialdata)}.}

\item{N0_treatment}{scalar. Number of observations of the historical treatment group.}

\item{N0_control}{scalar. Number of observations of the historical control group.}

\item{alternative}{character. The string specifying the alternative
hypothesis, must be one of \code{"greater"}, \code{"less"} or
\code{"two.sided"}.}

\item{N_impute}{scalar. Number of imputations for Monte Carlo simulation of
missing data.}

\item{number_mcmc}{scalar. Number of Monte Carlo Markov Chain draws in
sampling posterior.}

\item{prob_ha}{scalar. Probability of alternative hypothesis.}

\item{futility_prob}{scalar. Probability of stopping early for futility.}

\item{expected_success_prob}{scalar. Probability of stopping early for success.}

\item{discount_function}{character. If incorporating historical data, specify
the discount function. Currently supports the Weibull function
(\code{discount_function="weibull"}), the scaled-Weibull function
(\code{discount_function="scaledweibull"}), and the identity function
(\code{discount_function="identity"}). The scaled-Weibull discount function
scales the output of the Weibull CDF to have a max value of 1. The identity
discount function uses the posterior probability directly as the discount
weight. Default value is \code{"identity"}. See \code{\link{bdpnormal}} for
more details.}

\item{fix_alpha}{logical. Fix alpha at alpha_max? Default value is FALSE.}

\item{alpha_max}{scalar. Maximum weight the discount function can apply.
Default is 1. For a two-arm trial, users may specify a vector of two values
where the first value is used to weight the historical treatment group and
the second value is used to weight the historical control group.}
}
\value{
a list of output for the bayesian trial for binomial count.
}
\description{
Simulation for binomial counts for Bayesian Adaptive trial with
 different inputs to control for power, sample size, type 1 error rate, etc.
}
