% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialBACT.R
\name{data_binomial}
\alias{data_binomial}
\title{Data file for binomial analysis}
\usage{
data_binomial(data, .data = NULL)
}
\arguments{
\item{data}{data frame. A data frame which provides patient id, treatment group, outcome
of the treatment and complete columns. An example file is available at
\code{data(binomialdata)}.}

\item{.data}{NULL. stores the binomial data for analysis, please do not fill it in.}
}
\value{
a list with binomial data file to be analyzed.
}
\description{
Wrapper function for data file in binomial analysis.
}
\examples{
data_binomial(data = binomialdata)
}
