% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{data_normal}
\alias{data_normal}
\title{Data file for normal analysis}
\usage{
data_normal(data, .data = NULL)
}
\arguments{
\item{data}{data frame. A data frame which provides patient id, treatment group, outcome
of the treatment and complete columns. An example file is available at
\code{data(normaldata)}.}

\item{.data}{NULL. stores the normal data for analysis, please do not fill it in.}
}
\value{
a list with normal data file to be analyzed.
}
\description{
Wrapper function for data file in normal analysis.
}
\examples{
data_normal(data = normaldata)
}
