% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment.R
\name{enrollment}
\alias{enrollment}
\title{Simulating enrollment dates}
\usage{
enrollment(param, N_total, time = NULL)
}
\arguments{
\item{param}{a vector of lambda in poisson}

\item{N_total}{a numeric value of}

\item{time}{a vector of the same size as param indicating end of time when a specific lambda is used}

\item{family}{a factor value of either "poisson" or "binomial"}
}
\value{
a vector of enrollment dates
}
\description{
This function simulates enrollment dates using either poisson or binomial distribution
}
\examples{
enrollment(param = c(0.003, 0.7), 100, time = c(10, 400))



}
