% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{enrollment_rate}
\alias{enrollment_rate}
\title{Enrollment rate wrapper}
\usage{
enrollment_rate(lambda = NULL, time = NULL, data = NULL)
}
\arguments{
\item{lambda}{vector. Vector with different enrollment rate.}

\item{time}{vector. Vector with different cutoff time for lambda.}

\item{data}{NULL. stores the impute function}
}
\value{
a list with enrollment rate information
}
\description{
Wrapper function for enrollment rate
}
\examples{
enrollment_rate(lambda = c(0.3, 1), time = 25)
}
