% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{gamma_prior}
\alias{gamma_prior}
\title{Gamma prior for for control and treatment group}
\usage{
gamma_prior(a0 = 0.1, b0 = 0.1, .data = NULL)
}
\arguments{
\item{a0}{numeric. The shape parameter in the gamma distribution
(\code{beta(a0, b0)}).}

\item{b0}{numeric. The scale parameter in the beta distribution
(\code{beta(a0, b0)}).}

\item{.data}{NULL. stores the gamma prior rate, please do not fill it in.}
}
\value{
a list with vector of gamma rate for the gamma prior for treatment and control group.
}
\description{
Wrapper function for gamma prior \code{Gamma(a0, b0)}.
}
\examples{
gamma_prior(a0 = .1, b0 = .1)
}
