% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{impute}
\alias{impute}
\title{Imputation wrapper}
\usage{
impute(no_of_impute, number_mcmc, .data = NULL)
}
\arguments{
\item{no_of_impute}{integer. Number of Monte Carlo imputation for missing data}

\item{number_mcmc}{scalar. Number of Monte Carlo Markov Chain in sampling posterior.}

\item{.data}{NULL. stores the impute function}
}
\value{
a list with number of imputation
}
\description{
Wrapper function for no_of _impute
}
\examples{
impute(no_of_impute = 100, number_mcmc = 1000)
}
