% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialBACT.R
\name{looks}
\alias{looks}
\title{Interim look wrapper}
\usage{
looks(interim_look = NULL, data = NULL)
}
\arguments{
\item{interim_look}{vector. Vector with interim looks.}

\item{data}{NULL. stores the impute function}
}
\value{
a list with interim look information
}
