% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialBACT.R
\name{n_impute}
\alias{n_impute}
\title{Imputation wrapper}
\usage{
n_impute(no_of_impute, data = NULL)
}
\arguments{
\item{no_of_impute}{integer. Number of Monte Carlo imputation for missing data}

\item{data}{NULL. stores the impute function}
}
\value{
a list with number of imputation
}
