% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalBACT.R
\name{normal_analysis}
\alias{normal_analysis}
\title{Analyzing bayesian trial for normal mean data}
\usage{
normal_analysis(data = NULL, mu0_treatment = NULL,
  sd0_treatment = NULL, N0_treatment = NULL, mu0_control = NULL,
  sd0_control = NULL, N0_control = NULL, alternative = "greater",
  N_impute = 100, h0 = 0, number_mcmc = 10000, prob_ha = 0.95,
  futility_prob = 0.1, expected_success_prob = 0.9,
  discount_function = "identity", fix_alpha = FALSE, alpha_max = 1,
  weibull_scale = 0.135, weibull_shape = 3)
}
\arguments{
\item{data}{data frame. A data frame which provides patient id, treatment group, outcome
of the treatment and complete columns. An example file is available at
\code{data(normaldata)}.}

\item{mu0_treatment}{scalar. Mean of the historical treatment group.}

\item{sd0_treatment}{scalar. Standard deviation of the historical treatment group.}

\item{N0_treatment}{scalar. Number of observations of the historical treatment group.}

\item{mu0_control}{scalar. Mean of the historical control group.}

\item{sd0_control}{scalar. Standard deviation of the historical control group.}

\item{N0_control}{scalar. Number of observations of the historical control group.}

\item{alternative}{character. The string specifying the alternative
hypothesis, must be one of \code{"greater"}, \code{"less"} or
\code{"two.sided"}.}

\item{N_impute}{scalar. Number of imputations for Monte Carlo simulation of
missing data.}

\item{h0}{scalar. Threshold for comparing two mean values. Default is
\code{h0=0}.}

\item{number_mcmc}{scalar. Number of Monte Carlo Markov Chain draws in
sampling posterior.}

\item{prob_ha}{scalar. Probability of alternative hypothesis.}

\item{futility_prob}{scalar. Probability of stopping early for futility.}

\item{expected_success_prob}{scalar. Probability of stopping early for success.}

\item{discount_function}{character. If incorporating historical data, specify
the discount function. Currently supports the Weibull function
(\code{discount_function="weibull"}), the scaled-Weibull function
(\code{discount_function="scaledweibull"}), and the identity function
(\code{discount_function="identity"}). The scaled-Weibull discount function
scales the output of the Weibull CDF to have a max value of 1. The identity
discount function uses the posterior probability directly as the discount
weight. Default value is \code{"identity"}. See \code{\link{bdpnormal}} for
more details.}

\item{fix_alpha}{logical. Fix alpha at alpha_max? Default value is FALSE.}

\item{alpha_max}{scalar. Maximum weight the discount function can apply.
Default is 1. For a two-arm trial, users may specify a vector of two values
where the first value is used to weight the historical treatment group and
the second value is used to weight the historical control group.}

\item{weibull_scale}{scalar. Scale parameter of the Weibull discount function
used to compute alpha, the weight parameter of the historical data. Default
value is 0.135. For a two-arm trial, users may specify a vector of two
values where the first value is used to estimate the weight of the
historical treatment group and the second value is used to estimate the
weight of the historical control group. Not used when discount_function =
"identity".}

\item{weibull_shape}{scalar. Shape parameter of the Weibull discount function
used to compute alpha, the weight parameter of the historical data. Default
value is 3. For a two-arm trial, users may specify a vector of two values
where the first value is used to estimate the weight of the historical
treatment group and the second value is used to estimate the weight of the
historical control group. Not used when discount_function = "identity".}
}
\value{
a list of output for the bayesian trial for normal mean.
}
\description{
Function to analyze bayesian trial for normal mean data
 which allows early stopping and incorporation of historical data using
 the discount function approach
}
