% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalBACT.R
\name{normal_outcome}
\alias{normal_outcome}
\title{Parameters for treatment and control in normal case}
\usage{
normal_outcome(mu_control_true = NULL, sd_control_true = NULL,
  mu_treatment_true = NULL, sd_treatment_true = NULL, data = NULL)
}
\arguments{
\item{mu_control_true}{numeric. The mean for the control group.}

\item{sd_control_true}{numeric. The standard deviation for the control group.}

\item{mu_treatment_true}{numeric. The mean for the treatment group.}

\item{sd_treatment_true}{numeric. The standard deviation for the treatment group.}

\item{data}{NULL. stores the proportion of control and treatment, please do not fill it in.}
}
\value{
a list with proportion of control and treatment group.
}
\description{
Wrapper function for mean and standard deviation with normal outcome.
}
\examples{
normal_outcome(mu_control = 12, mu_treatment = 8, sd_treatment = 2.2, sd_control = 1.6)
}
