% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialBACT.R
\name{proportion}
\alias{proportion}
\title{Proportion of failure in control and treatment}
\usage{
proportion(p_control = NULL, p_treatment = NULL, data = NULL)
}
\arguments{
\item{p_control}{numeric. The proportion of failure in the control group, 0 < $p_control$ < 1.}

\item{p_treatment}{numeric. The proportion of failure in the treatment group, 0 < $p_treatment$ < 1.}

\item{data}{NULL. stores the proportion of control and treatment.}
}
\value{
a list with proportion of control and treatment group.
}
