% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{randomization}
\alias{randomization}
\title{Randomization scheme}
\usage{
randomization(N_total, block = 2, scheme = c(1, 1))
}
\arguments{
\item{N_total}{a numeric value of total sample size for randomization scheme}

\item{block}{a integer value of the block size for randomization, it needs to be a multiple of the sum of scheme}

\item{scheme}{a numeric vector of the randomization scheme \code{c(control, treatment)}}
}
\value{
the randomization scheme with 0, 1 for control and treatment
}
\description{
Implements randomization scheme for control and treatment with different randomization ratios and block
}
\examples{
#implementing randomization scheme for control to treatment with 1:1.5 randomization ratio
randomization(100, block = 5, c(2, 3))
#randomization scheme with 2:1 for control to treatment
randomization(70, block = 9, c(2, 1))

}
