% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialBACT.R
\name{randomize}
\alias{randomize}
\title{Posterior wrapper}
\usage{
randomize(block_size, randomization_ratio, data = NULL)
}
\arguments{
\item{data}{NULL. stores the randomization scheme function}
}
\value{
a list with
}
\description{
Wrapper function for the posterior function
}
