% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialBACT.R
\name{sample.size}
\alias{sample.size}
\title{Sample size wrapper}
\usage{
sample.size(sample.size, end.of.study, data)
}
\arguments{
\item{sample.size}{integer. The number of sample size needed.}

\item{end.of.study}{integer. The length of the study.}

\item{data}{NULL. stores the sample size and length of study.}
}
\value{
a list with sample size and length of the study.
}
