% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialBACT.R
\name{sample_size}
\alias{sample_size}
\title{Sample size wrapper}
\usage{
sample_size(sample.size, end.of.study, data = NULL)
}
\arguments{
\item{sample.size}{integer. The number of sample size needed.}

\item{end.of.study}{integer. The length of the study.}

\item{data}{NULL. stores the sample size and length of study.}
}
\value{
a list with sample size and length of the study.
}
\description{
Wrapper function for sample size and length of study.
}
\examples{
sample_size(sample.size = 300, end.of.study = 50)
}
