% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulation wrapper for binomial and normal.}
\usage{
simulate(input, no_of_sim = 10000, .data = NULL)
}
\arguments{
\item{input}{list. Input function for all inputs in binomial and normal .}

\item{no_of_sim}{numeric. Number of simulations to run}

\item{.data}{NULL. stores the proportion of control and treatment, please do not fill it in.}
}
\value{
a list with results of the simulation (power and type I error).
}
\description{
Wrapper function for complete binomial and normal function to compute power and type 1 error.
}
