% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalBACT.R
\name{BACTnormal}
\alias{BACTnormal}
\title{Complete normal wrapper function}
\usage{
BACTnormal(input, .data = NULL)
}
\arguments{
\item{input}{list. Input function for all normalBACT.}

\item{.data}{NULL. stores the proportion of control and treatment, please do not fill it in.}
}
\value{
a list with results of the clinical outcome.
}
\description{
Wrapper function for complete normal bayesCT function to compute power and type 1 error.
}
