% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{beta_prior}
\alias{beta_prior}
\title{Beta prior for for control and treatment group}
\usage{
beta_prior(a0 = 1, b0 = 1, .data = NULL)
}
\arguments{
\item{a0}{numeric. The first shape parameter in the beta distribution
(\code{beta(a0, b0)}).}

\item{b0}{numeric. The second shape parameter in the beta distribution
(\code{beta(a0, b0)}).}

\item{.data}{NULL. stores the beta prior rate, please do not fill it in.}
}
\value{
a list with vector of beta rate for the beta prior for treatment and control group.
}
\description{
Wrapper function for beta prior \code{beta(a0, b0)}.
}
\examples{
beta_prior(a0 = 1, b0 = 1)
}
