% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialBACT.R
\name{binomialBACT}
\alias{binomialBACT}
\title{Binomial counts for Bayesian Adaptive Trial}
\usage{
binomialBACT(p_treatment, p_control, N_total, lambda, lambda_time,
  analysis_at_enrollnumber, EndofStudy, weibull_scale, weibull_shape,
  discount_function = "identity", bpd_method = "fixed", block = 2,
  rand.ratio = c(1, 1), prop_loss_to_followup = 0.15, h0 = 0,
  futility_prob = 0.05, expected_success_prob = 0.9, prob_ha = 0.95,
  N_impute = 100)
}
\arguments{
\item{p_treatment}{scalar. Proportion of an event under the treatment group.}

\item{p_control}{scalar. Proportion of an event under the contrl group.}

\item{N_total}{scalar. Total sample size}

\item{lambda}{vector. Lambda for different enrollment rates across times.}

\item{lambda_time}{vector. Same size as lambda, denote times at lambda changes.}

\item{analysis_at_enrollnumber}{vector. Sample size for interim looks,
total size not included.}

\item{EndofStudy}{scalar. Length of the study.}

\item{weibull_scale}{vector. The weibull scale used in control and treatment group.}

\item{weibull_shape}{vector. The weibull shape used in control and treatment group.}

\item{discount_function}{character. The default is set to identity.
For more options look at bayesDP package.}

\item{bpd_method}{character. The default is set to fixed.
For more options look at bayesDP package.}

\item{block}{scalar. Block size for randomization to be implemented.}

\item{rand.ratio}{vector. Randomization ratio for control to treatment.
Integer values mapping the size of the block.}

\item{prop_loss_to_followup}{scalar. Proportion of loss in follow up.}

\item{h0}{scalar. treshold for comparing two proportions. default at 0.}

\item{futility_prob}{scalar. Type 1 error rate.}

\item{expected_success_prob}{scalar. Power of the study.}

\item{prob_ha}{scalar. Probability of alternative hypothesis.}

\item{N_impute}{scalar. Number of imputation for monte carlo simulation.}
}
\value{
a list of output
}
