% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialBACT.R
\name{binomialBACT}
\alias{binomialBACT}
\title{Binomial counts for Bayesian Adaptive Trials}
\usage{
binomialBACT(p_treatment, p_control = NULL, y0_treatment = NULL,
  N0_treatment = NULL, y0_control = NULL, N0_control = NULL,
  discount_function = "identity", N_total, lambda, lambda_time,
  interim_look = NULL, EndofStudy, prior = c(1, 1), block = 2,
  rand_ratio = c(1, 1), prop_loss_to_followup = 0.1,
  alternative = "two-sided", h0 = 0, futility_prob = 0.05,
  expected_success_prob = 0.9, prob_ha = 0.95, N_impute = 1000,
  number_mcmc = 1000)
}
\arguments{
\item{p_treatment}{scalar. Proportion of events under the treatment arm.}

\item{p_control}{scalar. Proportion of events under the control arm.}

\item{y0_treatment}{scalar. Number of events for the historical treatment
arm.}

\item{N0_treatment}{scalar. Sample size of the historical treatment arm.}

\item{y0_control}{scalar. Number of events for the historical control arm.}

\item{N0_control}{scalar. Sample size of the historical control arm.}

\item{discount_function}{character. Specify the discount function to use.
Currently supports the Weibull function
(\code{discount_function="weibull"}), the scaled-Weibull function
(\code{discount_function="scaledweibull"}), and the identity function
(\code{discount_function="identity"}). The scaled-Weibull discount function
scales the output of the Weibull CDF to have a max value of 1. The identity
discount function uses the posterior probability directly as the discount
weight. Default value is \code{"identity"}.}

\item{N_total}{scalar. Total sample size.}

\item{lambda}{vector. Lambda for different enrollment rates across times.}

\item{lambda_time}{vector. Same size as lambda, denote times at lambda changes.}

\item{interim_look}{vector. Sample size for interim looks. Note: the overall trial size should not be included.}

\item{EndofStudy}{scalar. Length of the study.}

\item{prior}{vector. Prior value of beta rate, beta(a0, b0).}

\item{block}{scalar. Block size for randomization to be implemented.}

\item{rand_ratio}{vector. Randomization allocation for control to treatment.
Integer values mapping the size of the block.}

\item{prop_loss_to_followup}{scalar. Proportion of subjects lost to follow-up.}

\item{alternative}{character. The string specifying the alternative hypothesis, must be one
of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}.}

\item{h0}{scalar. Treshold for comparing two proportions. Default is \code{h0=0}.}

\item{futility_prob}{scalar. Type I error rate.}

\item{expected_success_prob}{scalar. Power of the study.}

\item{prob_ha}{scalar. Probability of alternative hypothesis.}

\item{N_impute}{scalar. Number of imputations for Monte Carlo simulation of missing data.}

\item{number_mcmc}{scalar. Number of Monte Carlo Markov Chain draws in sampling posterior.}
}
\value{
a list of output
}
\description{
Simulation for binomial counts for Bayesian Adaptive trial with
  different inputs to control for power, sample size, type 1 error rate, etc.
}
\examples{
binomialBACT(p_control = 0.12, p_treatment = 0.10,
             y0_treatment = 8, N0_treatment = 90,
             y0_control = 13, N0_control = 95,
             N_total = 300, N_impute = 100,
             lambda = c(0.3, 1), lambda_time = c(25),
             interim_look = c(210, 240, 270),
             EndofStudy = 50)

}
