% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{data_normal}
\alias{data_normal}
\title{Data file for continuous (normally distributed) data analysis}
\usage{
data_normal(treatment, outcome, complete, .data = NULL)
}
\arguments{
\item{treatment}{vector. Treatment assignment for patients, 1 for treatment
group and 0 for control group}

\item{outcome}{vector. Normal outcome of the trial.}

\item{complete}{vector. Similar length as treatment and outcome variable, 1
for complete outcome, 0 for loss to follow up. If complete is not provided,
the dataset is assumed to be complete.}

\item{.data}{NULL. Stores the normal data for analysis. Should not be edited
by the user.}
}
\value{
a list with treatment, outcome and loss to follow up vector with
  normal outcome.
}
\description{
Wrapper function for data file in normal analysis.
}
