% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment.R
\name{enrollment}
\alias{enrollment}
\title{Simulating enrollment dates}
\usage{
enrollment(param, N_total, time = NULL)
}
\arguments{
\item{param}{vector. Lambda values for Poisson distribution.}

\item{N_total}{integer. Value of total sample size.}

\item{time}{vector. Knots (of \code{length(param)} - 1) indicating end of
time when a specific lambda is used.}
}
\value{
A vector of enrollment times (from time of first patient enrollment)
  in days.
}
\description{
This function simulates enrollment dates using either Poisson
  distribution
}
\examples{
enrollment(param = c(0.003, 0.7), 100, time = 10)

enrollment(param = c(0.3, 0.5, 0.9, 1.2, 2.1), 200, c(20, 30, 40, 60))

}
