% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{enrollment_rate}
\alias{enrollment_rate}
\title{Enrollment rate wrapper}
\usage{
enrollment_rate(lambda = 0.3, time = NULL, .data = NULL)
}
\arguments{
\item{lambda}{vector. Vector with different enrollment rate parameters.}

\item{time}{vector. Vector with different cut-off times for lambda.}

\item{.data}{NULL. This should not be changed by the user.}
}
\value{
a list with enrollment rate information
}
\description{
Wrapper function for enrollment rate.
}
\examples{
enrollment_rate(lambda = c(0.3, 1), time = 25)
}
