% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{hypothesis}
\alias{hypothesis}
\title{Hypothesis wrapper}
\usage{
hypothesis(delta, futility_prob, prob_ha, expected_success_prob, data = NULL)
}
\arguments{
\item{delta}{numeric. threshold set for margin in null hypothesis}

\item{futility_prob}{numeric. probability of futility}

\item{prob_ha}{numeric. posterior probability of accepting alternative.}

\item{expected_success_prob}{numeric. probability of expecting success.}

\item{data}{NULL. stores the randomization scheme function}
}
\value{
a list with information of hypothesis testing (threshold, futility probability,
probability of alternative and probability of expected success.)
}
\description{
Wrapper function for the hypothesis in the trial
}
\examples{
hypothesis(delta = 0, futility_prob = 0.05, prob_ha = 0.95, expected_success_prob = 0.90)
hypothesis(delta= 0.2, futility_prob = 0.1, prob_ha = 0.975, expected_success_prob = 0.80)
}
