% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{impute}
\alias{impute}
\title{Imputation wrapper}
\usage{
impute(no_of_impute = 10000, number_mcmc = 10000, .data = NULL)
}
\arguments{
\item{no_of_impute}{integer. Number of Monte Carlo imputation for missing
data.}

\item{number_mcmc}{scalar. Number of Monte Carlo Markov Chain draws from
posterior distribution.}

\item{.data}{NULL. This should not be changed by the user.}
}
\value{
a list with number of imputation
}
\description{
Wrapper function for no_of_impute.
}
\examples{
impute(no_of_impute = 100, number_mcmc = 1000)
}
