% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal_outcome}
\alias{normal_outcome}
\title{Parameters for treatment and control in normal case}
\usage{
normal_outcome(
  mu_control = NULL,
  sd_control = NULL,
  mu_treatment = NULL,
  sd_treatment = NULL,
  data = NULL
)
}
\arguments{
\item{mu_control}{numeric. The mean for the control group.}

\item{sd_control}{numeric. The standard deviation for the control group.}

\item{mu_treatment}{numeric. The mean for the treatment group.}

\item{sd_treatment}{numeric. The standard deviation for the treatment group.}

\item{data}{NULL. stores the proportion of control and treatment, please do not fill it in.}
}
\value{
a list with proportion of control and treatment group.
}
\description{
Wrapper function for mean and standard deviation with normal outcome.
}
\examples{
normal_outcome(mu_control = 12, mu_treatment = 8, sd_treatment = 2.2, sd_control = 1.6)
}
