% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaldata.R
\docType{data}
\name{normaldata}
\alias{normaldata}
\title{Gaussian dataset for analyzing adaptive Bayesian trials}
\format{
A data frame with 300 rows and 4 variables:
\describe{
  \item{id}{Patient ID in the trial}
  \item{treatment}{treatment assignment for patients, 1 for treatment group
  0 for control group}
  \item{outcome}{continuous outcome of the trial (gaussian distribution)}
  \item{complete}{1 for complete outcome, 0 for loss to follow up}
}
}
\usage{
data(normaldata)
}
\description{
A dataset containing the results of 300 patients with continuous (normal) outcome,
the dataset is filled with loss to follow up.
}
\examples{
data(normaldata)
}
\keyword{dataset}
